+++ src/Entities/Transaction/TransactionCreateItemWithPrice.php
@@ -11,6 +11,9 @@ declare(strict_types=1);
 
 namespace Paddle\SDK\Entities\Transaction;
 
+/**
+ * @deprecated Replaced by \Paddle\SDK\Resources\Transactions\Operations\Create\TransactionCreateItemWithPrice
+ */
 class TransactionCreateItemWithPrice
 {
     public function __construct(