+++ src/Entities/Transaction/TransactionUpdateTransactionItem.php
@@ -11,6 +11,9 @@ declare(strict_types=1);
 
 namespace Paddle\SDK\Entities\Transaction;
 
+/**
+ * @deprecated Replaced by \Paddle\SDK\Resources\Transactions\Operations\Update\TransactionUpdateItem
+ */
 class TransactionUpdateTransactionItem
 {
     public function __construct(